Api done

+++ API/Api.Application/Presistance/IGenericRepository.cs
@@ -6,7 +6,7 @@ namespace Api.Application.Presistance
 {
     public interface IGenericRepository<T> where T : class
     {
-        Task<T> Get(string tablename, int id);
+        Task<T> Get(string tablename, long id);
         Task<IEnumerable<T>> GetAll(string tablename);
         Task<T> Add(T entity);
         Task<bool> Exists(int id);
